Move _handle_attendance_tracking_role() to save_fn
